+++ src/components/index.ts
@@ -16,5 +16,6 @@ export { default as ResizeCropCard } from './ResizeCropCard';
 export { default as TemplateSelectionCard } from './TemplateSelectionCard';
 export { default as TemplateProcessingTab } from './TemplateProcessingTab';
 export { default as CustomProcessingTab } from './CustomProcessingTab';
-export { default as AdvancedRenameTab } from './AdvancedRenameTab'; // Add this line
+export { default as AdvancedRenameTab } from './AdvancedRenameTab';
 export { default as ErrorBoundary } from './ErrorBoundary';
+export * from './QualityImprovementCard';